Add comprehensive tests for cache joins and SQL functions
- Introduced `test_cache_join.jl` to validate the behavior of _cache_join in CTE contexts, ensuring correct join paths and data retrieval.
- Created `test_sql_functions.jl` to cover a wide range of SQL functions including aggregates, string manipulations, mathematical operations, and conditional logic.
- Implemented tests for advanced scenarios such as nested case statements, date functions, and complex filtering using Q objects.
- Ensured all tests validate expected outputs and handle edge cases effectively.
