+++ src/QueryBuilder.jl
@@ -51,7 +51,9 @@ include("querybuilder/ctes.jl")
 #
 export OP
 export Q, Qor
-export Sum, Avg, Count, Max, Min, When, F
+export Sum, Avg, Count, Max, Min, When, F, Case, Cast, Concat, Extract, To_char, Value
+export Coalesce, Greatest, Least, Lower, Upper, Length, Abs, Round, NullIf, Replace, Trim, LTrim, RTrim
+export Floor, Ceil, Sqrt, Exp, Ln, Power, Mod
 export With
