+++ src/models/fields.jl
@@ -1,0 +1,2380 @@
+
+struct sIDField <: PormGField
+  verbose_name::Union{String, Nothing}
+  primary_key::Bool
+  auto_increment::Bool
+  unique::Bool
+  blank::Bool
+  null::Bool
+  db_index::Bool
+  default::Union{Int64, Nothing}
+  editable::Bool
+  type::String
+  formater::Function
+  generated::Bool  # New field to indicate GENERATED ... AS IDENTITY
+  generated_always::Bool # New field to indicate GENERATED ALWAYS AS IDENTITY
+end
+
+"""
+    IDField(; kwargs...)
+
+A field type for auto-incrementing integer primary keys, equivalent to PostgreSQL's BIGSERIAL or GENERATED AS IDENTITY columns.
+
+The `IDField` is typically used as the primary key for models and automatically generates unique integer values for each record. It maps to a PostgreSQL BIGINT column with auto-increment capabilities.
+
+# Keyword Arguments
+- `verbose_name::Union{String, Nothing} = nothing`: A human-readable name for the field
+- `primary_key::Bool = true`: Whether this field is the primary key for the table
+- `auto_increment::Bool = true`: Whether the field should auto-increment (generate values automatically)
+- `unique::Bool = true`: Whether values in this field must be unique across all records
+- `blank::Bool = false`: Whether the field can be left blank in forms (not applicable for ID fields)
+- `null::Bool = false`: Whether the database column can store NULL values
+- `db_index::Bool = true`: Whether to create a database index on this field
+- `default::Union{Int64, Nothing} = nothing`: Default value for the field (rarely used with auto-increment)
+- `editable::Bool = false`: Whether the field should be editable in forms (typically false for ID fields)
+- `generated::Bool = true`: Whether to use PostgreSQL's GENERATED AS IDENTITY feature
+- `generated_always::Bool = false`: Whether to use GENERATED ALWAYS AS IDENTITY (stricter than regular GENERATED)
+
+# Database Mapping
+- **PostgreSQL Type**: BIGINT with GENERATED AS IDENTITY or GENERATED ALWAYS AS IDENTITY
+- **Auto-increment**: Supported through PostgreSQL's identity columns
+- **Index**: Automatically indexed as primary key
+
+# Examples
+
+Basic usage (most common):
+```julia
+User = Models.Model(
+    _id::PormGField = IDField()
+    name::PormGField = CharField(max_length=100)
+    email::PormGField = EmailField()
+)
+```
+
+Using GENERATED ALWAYS (stricter identity):
+```julia
+Order = Models.Model(
+    _id::PormGField = IDField(generated_always=true)
+    customer_id::PormGField = ForeignKey("Customer")
+    order_date::PormGField = DateTimeField()
+)
+```
+
+# Notes
+- The `IDField` is designed to be the primary key and should typically be the first field in your model
+- Values are automatically generated by the database, so you don't need to provide them when creating records
+- The field uses BIGINT type to support large ranges of ID values
+- When `generated_always=true`, the database will reject any attempts to manually insert ID values
+- This field type is PostgreSQL-specific and optimized for PormG's PostgreSQL backend
+
+# Validation
+- All boolean parameters are validated to ensure type safety
+- The `verbose_name` must be a String or nothing
+- The `default` value, if provided, must be convertible to Int64
+- Invalid parameters will trigger warnings but won't cause errors (they'll be ignored)
+"""
+function IDField(; kwargs...)
+  # List of accepted parameters
+  accepted = Set([
+      :verbose_name, :primary_key, :auto_increment, :unique, :blank, :null, :db_index, :default, :editable, :generated, :generated_always
+  ])
+  # Check for unexpected parameters
+  for (k, v) in kwargs
+      if !(k in accepted)
+          @warn "Unexpected parameter for IDField. It will be ignored." field="IDField" param=k value=v
+      end
+  end
+  # Extract parameters with defaults
+  verbose_name = get(kwargs, :verbose_name, nothing)
+  primary_key = get(kwargs, :primary_key, true)
+  auto_increment = get(kwargs, :auto_increment, true)
+  unique = get(kwargs, :unique, true)
+  blank = get(kwargs, :blank, false)
+  null = get(kwargs, :null, false)
+  db_index = get(kwargs, :db_index, true)
+  default = get(kwargs, :default, nothing)
+  editable = get(kwargs, :editable, false)
+  generated = get(kwargs, :generated, true)
+  generated_always = get(kwargs, :generated_always, false)
+
+  # Validate verbose_name
+  !(verbose_name isa Union{Nothing, String}) && throw(ArgumentError("The 'verbose_name' must be a String or nothing"))
+  # Validate other parameters
+  !(primary_key isa Bool) && throw(ArgumentError("The 'primary_key' must be a Boolean"))
+  !(auto_increment isa Bool) && throw(ArgumentError("The 'auto_increment' must be a Boolean"))
+  !(unique isa Bool) && throw(ArgumentError("The 'unique' must be a Boolean"))
+  !(blank isa Bool) && throw(ArgumentError("The 'blank' must be a Boolean"))
+  !(null isa Bool) && throw(ArgumentError("The 'null' must be a Boolean"))
+  !(db_index isa Bool) && throw(ArgumentError("The 'db_index' must be a Boolean"))
+  !(editable isa Bool) && throw(ArgumentError("The 'editable' must be a Boolean"))
+  !(generated isa Bool) && throw(ArgumentError("The 'generated' must be a Boolean"))
+  !(generated_always isa Bool) && throw(ArgumentError("The 'generated_always' must be a Boolean"))
+  # Validate default
+  default = validate_default(default, Union{Int64, Nothing}, "IDField", format2int64)
+  # Return the field instance
+  return sIDField(
+    verbose_name, primary_key, auto_increment, unique, blank, null, db_index, default, editable, "BIGINT", format_number_sql, generated, generated_always
+  )
+end
+
+mutable struct sForeignKey <: PormGField
+  verbose_name::Union{String, Nothing}
+  primary_key::Bool
+  unique::Bool
+  blank::Bool
+  null::Bool
+  db_index::Bool
+  default::Union{Int64, Nothing}
+  editable::Bool
+  to::Union{String, PormGModel, Nothing}
+  pk_field::Union{String, Symbol, Nothing}
+  on_delete::Union{Function, Nothing}
+  on_update::Union{String, Nothing}
+  deferrable::Bool
+  how::Union{String, Nothing}  # INNER JOIN, LEFT JOIN, RIGHT JOIN, FULL JOIN used in _build_row_join
+  related_name::Union{String, Nothing}
+  type::String
+  formater::Function
+  db_constraint::Bool
+  initially_deferred::Bool
+end
+
+"""
+    ForeignKey(to::Union{String, PormGModel}; kwargs...)
+
+A field that creates a many-to-one relationship to another model, similar to Django's ForeignKey.
+
+The `ForeignKey` field represents a relationship where many records in the current model can reference a single record in the target model. It creates a foreign key constraint in the database and enables efficient querying of related data.
+
+# Required Arguments
+- `to::Union{String, PormGModel}`: The target model that this field references. Can be either:
+  - A string with the model name (e.g., "User", "Category")  
+  - A direct reference to a PormGModel instance
+
+# Keyword Arguments
+- `verbose_name::Union{String, Nothing} = nothing`: A human-readable name for the field
+- `primary_key::Bool = false`: Whether this field is the primary key (rarely used with ForeignKey)
+- `unique::Bool = false`: Whether values must be unique (creates a one-to-one relationship if true)
+- `blank::Bool = false`: Whether the field can be left blank in forms
+- `null::Bool = false`: Whether the database column can store NULL values
+- `db_index::Bool = true`: Whether to create a database index on this field (recommended for performance)
+- `default::Union{Int64, Nothing} = nothing`: Default value for the field (ID of the referenced record)
+- `editable::Bool = false`: Whether the field should be editable in forms
+- `pk_field::Union{String, Symbol, Nothing} = nothing`: Which field in the target model to reference (defaults to primary key)
+- `on_delete::Union{Function, String, Nothing} = nothing`: Action when the referenced object is deleted
+- `on_update::Union{String, Nothing} = nothing`: Action when the referenced object's key is updated
+- `deferrable::Bool = false`: Whether the constraint check can be deferred until transaction commit
+- `initially_deferred::Bool = false`: Whether constraint checking is initially deferred
+- `how::Union{String, Nothing} = nothing`: Join type for queries ("INNER JOIN", "LEFT JOIN", etc.)
+- `related_name::Union{String, Nothing} = nothing`: Name for the reverse relation
+- `db_constraint::Bool = true`: Whether to create a database foreign key constraint
+
+# Database Mapping
+- **PostgreSQL Type**: BIGINT with foreign key constraint
+- **Constraint**: Creates `FOREIGN KEY` constraint linking to target table
+- **Index**: Automatically indexed for query performance
+
+# On Delete Options
+The `on_delete` parameter controls what happens when the referenced object is deleted:
+- `CASCADE`: Delete this object when referenced object is deleted
+- `RESTRICT`: Prevent deletion of referenced object if this object exists
+- `SET_NULL`: Set this field to NULL (requires `null=true`)
+- `SET_DEFAULT`: Set this field to its default value (requires `default` to be set)
+- `SET`: Set this field to a specific value
+- `PROTECT`: Raise an error to prevent deletion
+- `DO_NOTHING`: Take no action (may cause database integrity errors)
+
+# Examples
+
+Basic foreign key relationship:
+```julia
+Article = Models.Model(
+    _id = IDField()
+    title = CharField(max_length=200)
+    author = ForeignKey("User")
+    category = ForeignKey("Category", on_delete=CASCADE)
+)
+```
+
+Foreign key allowing NULL values:
+```julia
+Product = Models.Model(
+    _id = IDField()
+    _id = IDField()
+    name = CharField(max_length=100)
+    category = ForeignKey("Category", null=true, blank=true, on_delete=SET_NULL)
+)
+```
+
+Multiple foreign keys to same model (requires related_name):
+```julia
+Message = Models.Model(
+    _id = IDField()
+    sender = ForeignKey("User", related_name="sent_messages")
+    recipient = ForeignKey("User", related_name="received_messages")
+    content = TextField()
+)
+```
+
+# Related Names and Reverse Relations
+- If `related_name` is not specified, PormG automatically generates one
+- When multiple ForeignKeys point to the same model, `related_name` must be explicitly set
+- The related name allows querying from the target model back to this model
+- If you can't remember the related name, you can type `your_query.objects.related_objects` or `your_model.related_objects` to see all related names
+
+# Database Constraints
+- When `db_constraint=true` (default), creates actual foreign key constraints in PostgreSQL
+- When `db_constraint=false`, no database constraint is created (useful for legacy databases)
+- Database constraints ensure referential integrity but may impact performance
+
+# Validation
+- The `to` parameter must be a valid model name or PormGModel instance
+- All boolean parameters are validated for type safety
+- The `on_delete` parameter is validated against allowed values
+- Invalid parameters trigger warnings but don't cause errors
+
+# Notes
+- The field stores the primary key value of the referenced object
+- Uses BIGINT type to match IDField primary keys
+- Supports deferred constraint checking for complex transactions
+- Compatible with PostgreSQL's foreign key features
+
+# See Also
+- Django's ForeignKey documentation for conceptual understanding
+"""
+function ForeignKey(to::Union{String, PormGModel}; kwargs...)
+  # List of accepted parameters
+  accepted = Set([
+      :verbose_name, :primary_key, :unique, :blank, :null, :db_index, :default, :editable, :pk_field, :on_delete, :on_update, :deferrable, :initially_deferred, :how, :related_name, :db_constraint
+  ])
+  # Check for unexpected parameters
+  for (k, v) in kwargs
+      if !(k in accepted)
+          @warn "Unexpected parameter for ForeignKey. It will be ignored." field="ForeignKey" param=k value=v
+      end
+  end
+  # Extract parameters with defaults
+  verbose_name = get(kwargs, :verbose_name, nothing)
+  primary_key = get(kwargs, :primary_key, false)
+  unique = get(kwargs, :unique, false)
+  blank = get(kwargs, :blank, false)
+  null = get(kwargs, :null, false)
+  db_index = get(kwargs, :db_index, true)
+  default = get(kwargs, :default, nothing)
+  editable = get(kwargs, :editable, false)
+  pk_field = get(kwargs, :pk_field, nothing)
+  on_delete = get(kwargs, :on_delete, nothing)
+  on_update = get(kwargs, :on_update, nothing)
+  deferrable = get(kwargs, :deferrable, false)
+  initially_deferred = get(kwargs, :initially_deferred, false)
+  how = get(kwargs, :how, nothing)
+  related_name = get(kwargs, :related_name, nothing)
+  db_constraint = get(kwargs, :db_constraint, true)
+
+  # Validate 'to' parameter
+  !(to isa Union{String, PormGModel}) && throw(ArgumentError("The 'to' parameter must be a String or PormGModel"))
+
+  # Validate verbose_name
+  !(verbose_name isa Union{Nothing, String}) && throw(ArgumentError("The 'verbose_name' must be a String or nothing"))
+
+  # Validate boolean parameters
+  !(primary_key isa Bool) && throw(ArgumentError("The 'primary_key' must be a Boolean"))
+  !(unique isa Bool) && throw(ArgumentError("The 'unique' must be a Boolean"))
+  !(blank isa Bool) && throw(ArgumentError("The 'blank' must be a Boolean"))
+  !(null isa Bool) && throw(ArgumentError("The 'null' must be a Boolean"))
+  !(db_index isa Bool) && throw(ArgumentError("The 'db_index' must be a Boolean"))
+  !(editable isa Bool) && throw(ArgumentError("The 'editable' must be a Boolean"))
+  !(deferrable isa Bool) && throw(ArgumentError("The 'deferrable' must be a Boolean"))
+  !(initially_deferred isa Bool) && throw(ArgumentError("The 'initially_deferred' must be a Boolean"))
+
+  # Validate default
+  default = validate_default(default, Union{Int64, Nothing}, "ForeignKey", format2int64)
+
+  # Validate optional string parameters
+  !(pk_field isa Union{Nothing, AbstractString, Symbol}) &&
+    throw(ArgumentError("The 'pk_field' must be a String, Symbol, or nothing"))
+  on_delete = _get_on_delete_mode(on_delete)
+  !(on_update isa Union{Nothing, AbstractString}) && throw(ArgumentError("The 'on_update' must be a String or nothing"))
+  !(how isa Union{Nothing, AbstractString}) && throw(ArgumentError("The 'how' must be a String or nothing"))
+  !(related_name isa Union{Nothing, AbstractString}) && throw(ArgumentError("The 'related_name' must be a String or nothing"))
+  !(db_constraint isa Bool) && throw(ArgumentError("The 'db_constraint' must be a Boolean"))
+
+  # Resolve db_index based on db_constraint
+  db_index = db_index || !db_constraint 
+  pk_field = format_fild_name(pk_field)  
+
+  return sForeignKey(
+    verbose_name,
+    primary_key,
+    unique,
+    blank,
+    null,
+    db_index,
+    default,
+    editable,
+    to,
+    pk_field,
+    on_delete,
+    on_update,
+    deferrable,
+    how,
+    related_name,
+    "BIGINT",
+    format_number_sql,
+    db_constraint,
+    initially_deferred
+  )
+end
+
+function _get_on_delete_mode(on_delete::Nothing)
+  return nothing
+end
+function _get_on_delete_mode(on_delete::AbstractString)
+  on_delete = uppercase(on_delete)
+  if contains(on_delete, "CASCADE")
+    return CASCADE
+  elseif contains(on_delete, "RESTRICT")
+    return RESTRICT
+  elseif contains(on_delete, "SET_NULL")
+    return SET_NULL
+  elseif contains(on_delete, "SET_DEFAULT")
+    return SET_DEFAULT
+  elseif contains(on_delete, "SET")
+    return SET
+  elseif contains(on_delete, "PROTECT")
+    return PROTECT
+  elseif contains(on_delete, "DO_NOTHING")
+    return DO_NOTHING
+  else
+    throw(ArgumentError("The on_delete parameter must be CASCADE, RESTRICT, SET_NULL, SET_DEFAULT, SET, DO_NOTHING or PROTECT"))
+  end
+end
+function _get_on_delete_mode(on_delete::Function)
+  # check if the function is one of the valid functions
+  check_function = on_delete |> string |> uppercase
+  if !(check_function in ["CASCADE", "RESTRICT", "SET_NULL", "SET_DEFAULT", "SET", "DO_NOTHING", "PROTECT"])
+    throw(ArgumentError("The on_delete parameter must be CASCADE, RESTRICT, SET_NULL, SET_DEFAULT, SET, DO_NOTHING or PROTECT"))
+  end
+  return on_delete
+end
+
+
+mutable struct sOneToOneField <: PormGField
+  unique::Bool
+  verbose_name::Union{String, Nothing}
+  primary_key::Bool
+  blank::Bool
+  null::Bool
+  db_index::Bool
+  default::Union{Int64, Nothing}
+  editable::Bool
+  to::Union{String, PormGModel, Nothing}
+  pk_field::Union{String, Symbol, Nothing}
+  on_delete::Union{Function, Nothing}
+  on_update::Union{String, Nothing}
+  deferrable::Bool
+  how::Union{String, Nothing}  # INNER JOIN, LEFT JOIN, RIGHT JOIN, FULL JOIN used in _build_row_join
+  related_name::Union{String, Nothing}
+  type::String
+  formater::Function
+  db_constraint::Bool
+  initially_deferred::Bool
+end
+
+"""
+    OneToOneField(to::Union{String, PormGModel}; kwargs...)
+
+A field that creates a one-to-one relationship to another model, similar to Django's OneToOneField.
+
+The `OneToOneField` represents a strict one-to-one relationship where each record in the current model corresponds to exactly one record in the target model, and vice versa. It's essentially a ForeignKey with a unique constraint that ensures no two records can reference the same target record.
+
+# Required Arguments
+- `to::Union{String, PormGModel}`: The target model that this field references. Can be either:
+  - A string with the model name (e.g., "UserProfile", "Settings")  
+  - A direct reference to a PormGModel instance
+
+# Keyword Arguments
+- `verbose_name::Union{String, Nothing} = nothing`: A human-readable name for the field
+- `primary_key::Bool = false`: Whether this field is the primary key (rarely used with OneToOneField)
+- `unique::Bool = true`: Whether values must be unique (always true for one-to-one relationships)
+- `blank::Bool = false`: Whether the field can be left blank in forms
+- `null::Bool = false`: Whether the database column can store NULL values
+- `db_index::Bool = true`: Whether to create a database index on this field (recommended for performance)
+- `default::Union{Int64, Nothing} = nothing`: Default value for the field (ID of the referenced record)
+- `editable::Bool = false`: Whether the field should be editable in forms
+- `pk_field::Union{String, Symbol, Nothing} = nothing`: Which field in the target model to reference (defaults to primary key)
+- `on_delete::Union{Function, String, Nothing} = nothing`: Action when the referenced object is deleted
+- `on_update::Union{String, Nothing} = nothing`: Action when the referenced object's key is updated
+- `deferrable::Bool = false`: Whether the constraint check can be deferred until transaction commit
+- `initially_deferred::Bool = false`: Whether constraint checking is initially deferred
+- `how::Union{String, Nothing} = nothing`: Join type for queries ("INNER JOIN", "LEFT JOIN", etc.)
+- `related_name::Union{String, Nothing} = nothing`: Name for the reverse relation
+- `db_constraint::Bool = true`: Whether to create a database foreign key constraint
+
+# Database Mapping
+- **PostgreSQL Type**: BIGINT with unique foreign key constraint
+- **Constraint**: Creates `FOREIGN KEY` constraint with `UNIQUE` constraint
+- **Index**: Automatically indexed for query performance and uniqueness enforcement
+
+# One-to-One Relationship Characteristics
+- **Uniqueness**: Each target record can only be referenced by one record in the current model
+- **Bidirectional**: The relationship can be traversed in both directions
+- **Inheritance**: Often used to extend models without modifying the original table
+- **Profile Pattern**: Commonly used for user profiles, settings, or detailed information tables
+
+# On Delete Options
+The `on_delete` parameter controls what happens when the referenced object is deleted:
+- `CASCADE`: Delete this object when referenced object is deleted
+- `RESTRICT`: Prevent deletion of referenced object if this object exists
+- `SET_NULL`: Set this field to NULL (requires `null=true`)
+- `SET_DEFAULT`: Set this field to its default value (requires `default` to be set)
+- `SET`: Set this field to a specific value
+- `PROTECT`: Raise an error to prevent deletion
+- `DO_NOTHING`: Take no action (may cause database integrity errors)
+
+# Examples
+
+Basic one-to-one relationship (User Profile pattern):
+```julia
+User = Models.Model(
+    _id = IDField()
+    username = CharField(max_length=150, unique=true)
+    email = EmailField()
+)
+
+UserProfile = Models.Model(
+    _id = IDField()
+    user = OneToOneField("User", on_delete=CASCADE)
+    bio = TextField(blank=true)
+    avatar = ImageField(blank=true)
+    birth_date = DateField(null=true, blank=true)
+)
+```
+
+One-to-one with null values allowed:
+```julia
+Employee = Models.Model(
+    _id = IDField()
+    name = CharField(max_length=100)
+    department = CharField(max_length=50)
+)
+
+EmployeeSettings = Models.Model(
+    _id = IDField()
+    employee = OneToOneField("Employee", null=true, blank=true, on_delete=SET_NULL)
+    email_notifications = BooleanField(default=true)
+    theme_preference = CharField(max_length=20, default="light")
+)
+```
+
+Extending a model without modifying it:
+```julia
+Product = Models.Model(
+    _id = IDField()
+    name = CharField(max_length=200)
+    price = DecimalField(max_digits=10, decimal_places=2)
+)
+
+ProductDetails = Models.Model(
+    _id = IDField()
+    product = OneToOneField("Product", on_delete=CASCADE, related_name="details")
+    detailed_description = TextField()
+    technical_specs = TextField()
+    warranty_info = TextField()
+)
+```
+
+# Database Constraints vs. Unique ForeignKey
+OneToOneField is equivalent to:
+```julia
+# These are functionally identical:
+user = OneToOneField("User")
+user = ForeignKey("User", unique=true)
+```
+
+However, OneToOneField is more explicit about the intended relationship type and provides better semantic meaning.
+
+# Validation
+- The `to` parameter must be a valid model name or PormGModel instance
+- All boolean parameters are validated for type safety
+- The `on_delete` parameter is validated against allowed values
+- Uniqueness is automatically enforced at the database level
+- Invalid parameters trigger warnings but don't cause errors
+
+# See Also
+- Django's OneToOneField documentation for conceptual understanding
+- Database normalization principles for when to use one-to-one relationships
+"""
+function OneToOneField(to::Union{String, PormGModel}; kwargs...)
+  # List of accepted parameters
+  accepted = Set([
+      :verbose_name, :primary_key, :unique, :blank, :null, :db_index, :default, :editable, :pk_field, :on_delete, :on_update, :deferrable, :initially_deferred, :how, :related_name, :db_constraint
+  ])
+  # Check for unexpected parameters
+  for (k, v) in kwargs
+      if !(k in accepted)
+          @warn "Unexpected parameter for OneToOneField. It will be ignored." field="OneToOneField" param=k value=v
+      end
+  end
+  # Extract parameters with defaults
+  verbose_name = get(kwargs, :verbose_name, nothing)
+  primary_key = get(kwargs, :primary_key, false)
+  unique = get(kwargs, :unique, true)
+  blank = get(kwargs, :blank, false)
+  null = get(kwargs, :null, false)
+  db_index = get(kwargs, :db_index, true)
+  default = get(kwargs, :default, nothing)
+  editable = get(kwargs, :editable, false)
+  pk_field = get(kwargs, :pk_field, nothing)
+  on_delete = get(kwargs, :on_delete, nothing)
+  on_update = get(kwargs, :on_update, nothing)
+  deferrable = get(kwargs, :deferrable, false)
+  initially_deferred = get(kwargs, :initially_deferred, false)
+  how = get(kwargs, :how, nothing)
+  related_name = get(kwargs, :related_name, nothing)
+  db_constraint = get(kwargs, :db_constraint, true)
+
+  # Validate 'to' parameter
+  !(to isa Union{String, PormGModel}) && throw(ArgumentError("The 'to' parameter must be a String or PormGModel"))
+
+  # Validate verbose_name
+  !(verbose_name isa Union{Nothing, String}) && throw(ArgumentError("The 'verbose_name' must be a String or nothing"))
+
+  # Validate boolean parameters
+  !(primary_key isa Bool) && throw(ArgumentError("The 'primary_key' must be a Boolean"))
+  !(unique isa Bool) && throw(ArgumentError("The 'unique' must be a Boolean"))
+  !(blank isa Bool) && throw(ArgumentError("The 'blank' must be a Boolean"))
+  !(null isa Bool) && throw(ArgumentError("The 'null' must be a Boolean"))
+  !(db_index isa Bool) && throw(ArgumentError("The 'db_index' must be a Boolean"))
+  !(editable isa Bool) && throw(ArgumentError("The 'editable' must be a Boolean"))
+  !(deferrable isa Bool) && throw(ArgumentError("The 'deferrable' must be a Boolean"))
+  !(initially_deferred isa Bool) && throw(ArgumentError("The 'initially_deferred' must be a Boolean"))
+
+  # Validate default
+  default = validate_default(default, Union{Int64, Nothing}, "OneToOneField", format2int64)
+
+  # Validate optional string parameters
+  !(pk_field isa Union{Nothing, String, Symbol}) && throw(ArgumentError("The 'pk_field' must be a String, Symbol, or nothing"))
+  !(on_update isa Union{Nothing, AbstractString}) && throw(ArgumentError("The 'on_update' must be a String or nothing"))
+  !(how isa Union{Nothing, String}) && throw(ArgumentError("The 'how' must be a String or nothing"))
+  !(related_name isa Union{Nothing, String}) && throw(ArgumentError("The 'related_name' must be a String or nothing"))
+  !(db_constraint isa Bool) && throw(ArgumentError("The 'db_constraint' must be a Boolean"))
+
+  # Resolve on_delete using similar logic as ForeignKey
+  on_delete = _get_on_delete_mode(on_delete)
+
+  # Resolve db_index based on db_constraint
+  db_index = db_index || !db_constraint
+
+  return sOneToOneField(
+    unique,
+    verbose_name,
+    primary_key,
+    blank,
+    null,
+    db_index,
+    default,
+    editable,
+    to,
+    pk_field,
+    on_delete,
+    on_update,
+    deferrable,
+    how,
+    related_name,
+    "BIGINT",
+    format_number_sql,
+    db_constraint,
+    initially_deferred
+  )
+end
+
+mutable struct sAutoField <: PormGField
+  verbose_name::Union{String, Nothing}
+  primary_key::Bool
+  auto_increment::Bool
+  unique::Bool
+  blank::Bool
+  null::Bool
+  db_index::Bool
+  default::Union{Int64, Nothing}
+  editable::Bool
+  type::String
+  formater::Function
+end
+
+"""
+    AutoField(; kwargs...)
+
+A field type for auto-incrementing integer primary keys, equivalent to PostgreSQL's SERIAL columns.
+
+The `AutoField` is designed for auto-incrementing integer primary keys and automatically generates unique integer values for each record. Unlike `IDField` which uses BIGINT, `AutoField` uses INTEGER type and is suitable for applications that don't require the extended range of BIGINT values.
+
+# Keyword Arguments
+- `verbose_name::Union{String, Nothing} = nothing`: A human-readable name for the field
+- `primary_key::Bool = true`: Whether this field is the primary key for the table
+- `auto_increment::Bool = true`: Whether the field should auto-increment (generate values automatically)
+- `unique::Bool = true`: Whether values in this field must be unique across all records
+- `blank::Bool = false`: Whether the field can be left blank in forms (not applicable for auto fields)
+- `null::Bool = false`: Whether the database column can store NULL values
+- `db_index::Bool = false`: Whether to create a database index on this field (primary keys are automatically indexed)
+- `default::Union{Int64, Nothing} = nothing`: Default value for the field (rarely used with auto-increment)
+- `editable::Bool = false`: Whether the field should be editable in forms (typically false for auto fields)
+
+# Database Mapping
+- **PostgreSQL Type**: INTEGER with SERIAL auto-increment
+- **Auto-increment**: Supported through PostgreSQL's SERIAL type (sequence-based)
+- **Index**: Automatically indexed as primary key
+- **Range**: 32-bit signed integers (-2,147,483,648 to 2,147,483,647)
+
+# AutoField vs IDField Comparison
+| Feature | AutoField | IDField |
+|---------|-----------|---------|
+| **Database Type** | INTEGER (SERIAL) | BIGINT (BIGSERIAL/IDENTITY) |
+| **Range** | 32-bit (-2B to 2B) | 64-bit (-9Q to 9Q) |
+| **Storage** | 4 bytes | 8 bytes |
+| **Generation** | Sequence-based | Identity columns or sequence |
+| **Use Case** | Small to medium apps | Large-scale applications |
+
+
+# When to Use AutoField vs IDField
+
+**Use AutoField when:**
+- Building small to medium-sized applications
+- You don't expect more than ~2 billion records
+- Storage efficiency is important (4 bytes vs 8 bytes per ID)
+- Working with legacy systems that expect INTEGER primary keys
+- Building lookup tables, categories, or reference data
+
+**Use IDField when:**
+- Building large-scale applications with potential for massive growth
+- You need the extended range of 64-bit integers
+- Working with data warehouses or analytics platforms
+- Future-proofing against scale requirements
+- Using modern PostgreSQL features like identity columns
+"""
+function AutoField(; kwargs...)
+  # List of accepted parameters
+  accepted = Set([
+      :verbose_name, :primary_key, :auto_increment, :unique, :blank, :null, :db_index, :default, :editable
+  ])
+  # Check for unexpected parameters
+  for (k, v) in kwargs
+      if !(k in accepted)
+          @warn "Unexpected parameter for AutoField. It will be ignored." field="AutoField" param=k value=v
+      end
+  end
+  # Extract parameters with defaults
+  verbose_name = get(kwargs, :verbose_name, nothing)
+  primary_key = get(kwargs, :primary_key, true)
+  auto_increment = get(kwargs, :auto_increment, true)
+  unique = get(kwargs, :unique, true)
+  blank = get(kwargs, :blank, false)
+  null = get(kwargs, :null, false)
+  db_index = get(kwargs, :db_index, false)
+  default = get(kwargs, :default, nothing)
+  editable = get(kwargs, :editable, false)
+
+  # Validate verbose_name
+  !(verbose_name isa Union{Nothing, String}) && throw(ArgumentError("The 'verbose_name' must be a String or nothing"))
+  # Validate other parameters
+  !(primary_key isa Bool) && throw(ArgumentError("The 'primary_key' must be a Boolean"))
+  !(auto_increment isa Bool) && throw(ArgumentError("The 'auto_increment' must be a Boolean"))
+  !(unique isa Bool) && throw(ArgumentError("The 'unique' must be a Boolean"))
+  !(blank isa Bool) && throw(ArgumentError("The 'blank' must be a Boolean"))
+  !(null isa Bool) && throw(ArgumentError("The 'null' must be a Boolean"))
+  !(db_index isa Bool) && throw(ArgumentError("The 'db_index' must be a Boolean"))
+  !(editable isa Bool) && throw(ArgumentError("The 'editable' must be a Boolean"))
+  # Validate default
+  default = validate_default(default, Union{Int64, Nothing}, "AutoField", format2int64)
+  # Return the field instance
+
+  return sAutoField(verbose_name, primary_key, auto_increment, unique, blank, null, db_index, default, editable, "INTEGER", format_number_sql)
+end
+
+mutable struct sCharField <: PormGField
+  verbose_name::Union{String, Nothing}
+  primary_key::Bool
+  max_length::Int
+  unique::Bool
+  blank::Bool
+  null::Bool
+  db_index::Bool
+  db_column::Union{String, Nothing}
+  default::Union{String, Nothing}
+  editable::Bool
+  type::String
+  formater::Function
+  choices::Union{NTuple{N, Tuple{AbstractString, AbstractString}}, Nothing} where N
+end
+
+
+function parse_choices(choices_str::String)
+  # Parse a string into a tuple of tuples
+  # println(choices_str)
+  choices = ()
+  pattern = r"\(([^()]+)\)"
+  for m in eachmatch(pattern, choices_str)
+    inner = m.captures[1]
+    values = split(inner, ",")
+    if length(values) == 2
+      key = strip(values[1]) |> string
+      value = strip(values[2]) |> string
+      choices = (choices..., (key, value))
+    else
+      throw(ArgumentError("Invalid choices format"))
+    end
+  end
+  return choices
+end
+
+function count_just_strings(key_value::AbstractString)
+  count = 0
+  pattern = r"^\s*['\"](.*)['\"]\s*$"
+  for line in split(key_value, '\n')
+    if occursin(pattern, line)
+      count += 1
+    end
+  end
+  return count
+end
+
+function return_just_strings(key_value::AbstractString)
+  pattern = r"^\s*['\"](.*)['\"]\s*$"
+  m = match(pattern, key_value |> String)
+  if m !== nothing
+    return m.captures[1]
+  end  
+  return key_value
+end
+
+"""
+    CharField(; kwargs...)
+
+A field for storing short to medium-length strings, equivalent to PostgreSQL's VARCHAR columns.
+
+The `CharField` is the most commonly used field for storing textual data with a limited length. It maps to a PostgreSQL VARCHAR column and supports validation, indexing, choices, and various constraints. This field is ideal for names, titles, codes, and other string data with known maximum lengths.
+
+# Keyword Arguments
+- `verbose_name::Union{String, Nothing} = nothing`: A human-readable name for the field
+- `max_length::Int = 250`: Maximum number of characters allowed (1-255)
+- `unique::Bool = false`: Whether values in this field must be unique across all records
+- `blank::Bool = false`: Whether the field can be left blank in forms
+- `null::Bool = false`: Whether the database column can store NULL values
+- `db_index::Bool = false`: Whether to create a database index on this field
+- `db_column::Union{String, Nothing} = nothing`: Custom database column name (defaults to field name)
+- `default::Union{String, Nothing} = nothing`: Default value for the field
+- `choices::Union{NTuple{N, Tuple{AbstractString, AbstractString}}, Nothing} = nothing`: Restricted set of valid values
+- `editable::Bool = true`: Whether the field should be editable in forms
+
+# Length Constraints
+- **Minimum**: 1 character
+- **Maximum**: 255 characters
+- **Validation**: Automatically enforced at the field level
+- **Storage**: Efficient variable-length storage in PostgreSQL
+
+# Examples
+
+Basic string field:
+```julia
+User = Models.Model(
+    _id = IDField(),
+    username = CharField(max_length=150, unique=true),
+    first_name = CharField(max_length=50),
+    last_name = CharField(max_length=50)
+)
+```
+
+String field with choices (enumeration):
+```julia
+Order = Models.Model(
+    _id = IDField()
+    status = CharField(
+        max_length=20,
+        choices=(
+            ("1", "Pending"),
+            ("2", "Processing"),
+            ("3", "Shipped"),
+            ("4", "Delivered"),
+            ("5", "Cancelled")
+        ),
+        default="1"
+    )
+    customer_name = CharField(max_length=200)
+)
+```
+
+Optional field with custom database column:
+```julia
+Product = Models.Model(
+    _id = IDField(),
+    name = CharField(max_length=200),
+    sku = CharField(
+        max_length=50, 
+        unique=true, 
+        db_column="product_sku",
+        verbose_name="Stock Keeping Unit"
+    )
+    category = CharField(max_length=100, null=true, blank=true)
+)
+```
+
+Indexed field for performance:
+```julia
+Article = Models.Model(
+    _id = IDField(),
+    title = CharField(max_length=200, db_index=true),
+    slug = CharField(max_length=200, unique=true, db_index=true),
+    content = TextField()
+)
+```
+
+# Choices Feature
+The `choices` parameter allows you to restrict field values to a predefined set:
+
+```julia
+# Define choices as tuples of (value, display_name)
+priority_choices = (
+    ("low", "Low Priority"),
+    ("medium", "Medium Priority"),
+    ("high", "High Priority"),
+    ("urgent", "Urgent")
+)
+
+Task = Models.Model(
+    _id = IDField(),
+    title = CharField(max_length=200),
+    priority = CharField(max_length=10, choices=priority_choices, default="medium")
+)
+```
+
+**Choice Format Options:**
+1. **Tuple of Tuples**: `(("value1", "Display 1"), ("value2", "Display 2"))`
+2. **String Format**: `"(value1, Display 1)(value2, Display 2)"`
+
+# Default Values
+- **Static Default**: `default="some_value"`
+- **Must Match Choices**: If choices are specified, default must be one of the choice values
+- **Length Validation**: Default value must not exceed `max_length`
+
+# Database Column Naming
+- **Conventions**: Follow PostgreSQL naming conventions (lowercase, underscores)
+
+# CharField vs TextField
+| Feature | CharField | TextField |
+|---------|-----------|-----------|
+| **Length** | Limited (1-255) | Unlimited |
+| **Database Type** | VARCHAR | TEXT |
+| **Use Case** | Short strings | Long content |
+| **Indexing** | Efficient | Less efficient |
+| **Performance** | Fast queries | Slower for large content |
+
+# Migration Considerations
+- **Increasing Length**: Safe operation
+- **Decreasing Length**: Requires data validation
+- **Adding Choices**: Application-level change only
+- **Changing Column Name**: Use `db_column` parameter
+
+# Notes
+- The field uses VARCHAR type which is efficient for short to medium strings
+- Choices are validated at the Julia application level, not in the database
+- The `editable=true` default makes this field suitable for user input forms
+- Database indexing is optional but recommended for frequently queried fields
+- Compatible with PostgreSQL's text search and pattern matching features
+
+# See Also
+- `TextField` for unlimited length text content
+- `EmailField` for email address validation
+- Database design best practices for string field sizing
+"""
+function CharField(; kwargs...)
+  # List of accepted parameters
+  accepted = Set([
+      :verbose_name, :max_length, :unique, :blank, :null, :db_index, :db_column, :default, :choices, :editable
+  ])
+  # Check for unexpected parameters
+  for (k, v) in kwargs
+      if !(k in accepted)
+          @warn "Unexpected parameter for CharField. It will be ignored." field="CharField" param=k value=v
+      end
+  end
+  # Extract parameters with defaults
+  verbose_name = get(kwargs, :verbose_name, nothing)
+  max_length = get(kwargs, :max_length, 250)
+  unique = get(kwargs, :unique, false)
+  blank = get(kwargs, :blank, false)
+  null = get(kwargs, :null, false)
+  db_index = get(kwargs, :db_index, false)
+  db_column = get(kwargs, :db_column, nothing)
+  default = get(kwargs, :default, nothing)
+  choices = get(kwargs, :choices, nothing)
+  editable = get(kwargs, :editable, true)
+
+  !(verbose_name isa Union{Nothing, String}) && throw(ArgumentError("The verbose_name must be a String or nothing"))
+  max_length isa AbstractString && (max_length = parse(Int, max_length))
+  max_length isa Int || throw(ArgumentError("The max_length must be an integer"))
+  max_length > 255 && throw(ArgumentError("The max_length must be less than or equal to 255"))
+  max_length < 1 && throw(ArgumentError("The max_length must be greater than 1"))
+  default isa Int && (default = string(default))
+  if !(default isa Nothing) && !(default isa AbstractString) 
+    throw(ArgumentError("The default value must be a string, but got $(default) ($(typeof(default)))"))
+  end
+  if !(default isa Nothing) && length(default) > max_length
+    throw(ArgumentError("The default value exceeds the max_length, but got $(length(default)) and max_length is $(max_length)"))
+  end
+  !(unique isa Bool) && throw(ArgumentError("The unique must be a boolean"))
+  !(blank isa Bool) && throw(ArgumentError("The blank must be a boolean"))
+  !(null isa Bool) && throw(ArgumentError("The null must be a boolean"))
+  !(db_index isa Bool) && throw(ArgumentError("The db_index must be a boolean"))
+  !(db_column isa Union{Nothing, String}) && throw(ArgumentError("The db_column must be a string or nothing"))
+  !(editable isa Bool) && throw(ArgumentError("The editable must be a boolean"))  
+  if choices isa AbstractString
+    choices = parse_choices(choices)
+  elseif !(choices isa Union{Nothing, NTuple{N, Tuple{AbstractString, AbstractString}} where N })
+    println(choices)
+    println(choices |> typeof)
+    throw(ArgumentError("The 'choices' must be a String or Tuple{Tuple{String,String}}, but got $(choices) ($(typeof(choices)))"))
+  end
+  if choices !== nothing
+    for choice in choices
+      if !(choice[1] isa AbstractString)
+        throw(ArgumentError("Choice values must be strings"))
+      end
+      if count_just_strings(choice[1]) > max_length
+        throw(ArgumentError("Choices cannot exceed max_length"))
+      end
+    end
+    if default !== nothing
+      valid_defaults = choices isa Vector{String} ? return_just_strings(choices) : [return_just_strings(c[1]) for c in choices]
+      if !(default in valid_defaults)
+        throw(ArgumentError("The default value must be one of the choices"))
+      end
+    end
+  end
+  return sCharField(verbose_name, false, max_length, unique, blank, null, db_index, db_column, default, editable, "VARCHAR", format_text_sql, choices)
+end
+
+
+mutable struct sIntegerField <: PormGField
+  verbose_name::Union{String, Nothing}
+  primary_key::Bool
+  unique::Bool
+  blank::Bool
+  null::Bool
+  db_index::Bool
+  default::Union{Int64, Nothing}
+  editable::Bool
+  type::String
+  formater::Function
+end
+
+"""
+    IntegerField(; kwargs...)
+
+A field for storing 32-bit signed integers, equivalent to PostgreSQL's INTEGER columns.
+
+The `IntegerField` stores whole numbers within the 32-bit signed integer range (-2,147,483,648 to 2,147,483,647). It's ideal for counts, quantities, ratings, and other numeric data that doesn't require decimal places or extremely large values.
+
+# Keyword Arguments
+- `verbose_name::Union{String, Nothing} = nothing`: A human-readable name for the field
+- `unique::Bool = false`: Whether values in this field must be unique across all records
+- `blank::Bool = false`: Whether the field can be left blank in forms
+- `null::Bool = false`: Whether the database column can store NULL values
+- `db_index::Bool = false`: Whether to create a database index on this field
+- `default::Union{Int64, Nothing} = nothing`: Default value for the field
+- `editable::Bool = false`: Whether the field should be editable in forms
+
+# Examples
+
+Basic integer field:
+```julia
+Product = Models.Model(
+    _id = IDField(),
+    name = CharField(max_length=200),
+    quantity = IntegerField(default=0),
+    price_cents = IntegerField()  # Store price in cents to avoid decimals
+)
+```
+
+Integer field with constraints:
+```julia
+User = Models.Model(
+    _id = IDField(),
+    username = CharField(max_length=150, unique=true),
+    age = IntegerField(null=true, blank=true),
+    score = IntegerField(default=0, db_index=true)
+)
+```
+
+Rating system:
+```julia
+Review = Models.Model(
+    _id = IDField(),
+    product = ForeignKey("Product"),
+    rating = IntegerField(default=5),  # 1-5 star rating
+    helpful_votes = IntegerField(default=0)
+)
+```
+
+# Validation and Constraints
+- **Range**: Automatically validates within INTEGER bounds
+- **Type**: Accepts integers, numeric strings (converted automatically)
+- **Default**: Must be an integer or convertible to integer
+- **Null**: When `null=true`, accepts NULL values
+
+# Migration Considerations
+- **Range Changes**: Changing to BigIntegerField is safe
+- **Adding Constraints**: Adding uniqueness or indexes is safe
+- **Default Values**: Can be added or modified safely
+- **Null Constraints**: Removing null constraint requires data validation
+
+"""
+function IntegerField(; kwargs...)
+  # List of accepted parameters
+  accepted = Set([
+      :verbose_name, :unique, :blank, :null, :db_index, :default, :editable
+  ])
+  # Check for unexpected parameters
+  for (k, v) in kwargs
+      if !(k in accepted)
+          @warn "Unexpected parameter for IntegerField. It will be ignored." field="IntegerField" param=k value=v
+      end
+  end
+  # Extract parameters with defaults
+  verbose_name = get(kwargs, :verbose_name, nothing)
+  unique = get(kwargs, :unique, false)
+  blank = get(kwargs, :blank, false)
+  null = get(kwargs, :null, false)
+  db_index = get(kwargs, :db_index, false)
+  default = get(kwargs, :default, nothing)
+  editable = get(kwargs, :editable, false)
+
+  # Validate verbose_name
+  !(verbose_name isa Union{Nothing, String}) && throw(ArgumentError("The verbose_name must be a String or nothing"))
+  
+  # Validate default using validate_default
+  default = validate_default(default, Union{Int64, Nothing}, "IntegerField", format2int64)
+  
+  # Validate other parameters
+  !(unique isa Bool) && throw(ArgumentError("The 'unique' parameter must be a Boolean"))
+  !(blank isa Bool) && throw(ArgumentError("The 'blank' parameter must be a Booleadn"))
+  !(null isa Bool) && throw(ArgumentError("The 'null' parameter must be a Boolean"))
+  !(db_index isa Bool) && throw(ArgumentError("The 'db_index' parameter must be a Boolean"))
+  !(editable isa Bool) && throw(ArgumentError("The 'editable' parameter must be a Boolean"))
+  
+  return sIntegerField(
+    verbose_name,
+    false, # primary_key
+    unique,
+    blank,
+    null,
+    db_index,
+    default,
+    editable,
+    "INTEGER",
+    format_number_sql
+  )  
+end
+
+mutable struct sBigIntegerField <: PormGField
+  verbose_name::Union{String, Nothing}
+  primary_key::Bool
+  unique::Bool
+  blank::Bool
+  null::Bool
+  db_index::Bool
+  default::Union{Int64, Nothing}
+  editable::Bool
+  type::String
+  formater::Function
+end
+
+"""
+    BigIntegerField(; kwargs...)
+
+A field for storing 64-bit signed integers, equivalent to PostgreSQL's BIGINT columns.
+
+The `BigIntegerField` stores large whole numbers within the 64-bit signed integer range (-9,223,372,036,854,775,808 to 9,223,372,036,854,775,807). It's ideal for large identifiers, timestamps, population counts, and other numeric data requiring extended range beyond regular integers.
+
+# Keyword Arguments
+- `verbose_name::Union{String, Nothing} = nothing`: A human-readable name for the field
+- `unique::Bool = false`: Whether values in this field must be unique across all records
+- `blank::Bool = false`: Whether the field can be left blank in forms
+- `null::Bool = false`: Whether the database column can store NULL values
+- `db_index::Bool = false`: Whether to create a database index on this field
+- `default::Union{Int64, Nothing} = nothing`: Default value for the field
+- `editable::Bool = false`: Whether the field should be editable in forms
+
+# Database Mapping
+- **PostgreSQL Type**: BIGINT
+- **Storage**: 8 bytes per value
+- **Range**: -9,223,372,036,854,775,808 to 9,223,372,036,854,775,807
+- **Index**: Optional, recommended for frequently queried fields
+
+# Examples
+
+Large identifier field:
+```julia
+Analytics = Models.Model(
+    _id = IDField(),
+    user_id = BigIntegerField(db_index=true),
+    session_id = BigIntegerField(),
+    timestamp_ms = BigIntegerField()  # Unix timestamp in milliseconds
+)
+```
+
+Population and statistics:
+```julia
+Country = Models.Model(
+    _id = IDField(),
+    name = CharField(max_length=100),
+    population = BigIntegerField(null=true),
+    gdp_usd = BigIntegerField(null=true),  # GDP in USD cents
+    area_sq_meters = BigIntegerField()
+)
+```
+
+Large external identifiers:
+```julia
+SocialMedia = Models.Model(
+    _id = IDField(),
+    user = ForeignKey("User"),
+    twitter_id = BigIntegerField(unique=true, null=true),
+    facebook_id = BigIntegerField(unique=true, null=true),
+    follower_count = BigIntegerField(default=0)
+)
+```
+
+# Common Use Cases
+1. **Large Identifiers**: External API IDs, social media IDs
+2. **Timestamps**: Unix timestamps in milliseconds or microseconds
+3. **Population Data**: Country populations, large counts
+4. **Financial Data**: Large monetary values in smallest units
+5. **Scientific Data**: Large measurements, particle counts
+6. **Analytics**: Large user IDs, session identifiers
+
+# Migration Considerations
+- **From IntegerField**: Safe upgrade, no data loss
+- **To IntegerField**: Requires validation that all values fit in 32-bit range
+- **Index Changes**: Indexes will be recreated with new size
+- **Application Code**: May need updates if expecting different ranges
+"""
+function BigIntegerField(; kwargs...)
+  # List of accepted parameters
+  accepted = Set([
+      :verbose_name, :unique, :blank, :null, :db_index, :default, :editable
+  ])
+  # Check for unexpected parameters
+  for (k, v) in kwargs
+      if !(k in accepted)
+          @warn "Unexpected parameter for BigIntegerField. It will be ignored." field="BigIntegerField" param=k value=v
+      end
+  end
+  # Extract parameters with defaults
+  verbose_name = get(kwargs, :verbose_name, nothing)
+  unique = get(kwargs, :unique, false)
+  blank = get(kwargs, :blank, false)
+  null = get(kwargs, :null, false)
+  db_index = get(kwargs, :db_index, false)
+  default = get(kwargs, :default, nothing)
+  editable = get(kwargs, :editable, false)
+
+  # Validate verbose_name
+  !(verbose_name isa Union{Nothing, String}) && throw(ArgumentError("The verbose_name must be a String or nothing"))
+  
+  # Validate default using validate_default
+  default = validate_default(default, Union{Int64, Nothing}, "BigIntegerField", format2int64)
+  
+  # Validate other parameters
+  !(unique isa Bool) && throw(ArgumentError("The 'unique' parameter must be a Boolean"))
+  !(blank isa Bool) && throw(ArgumentError("The 'blank' parameter must be a Boolean"))
+  !(null isa Bool) && throw(ArgumentError("The 'null' parameter must be a Boolean"))
+  !(db_index isa Bool) && throw(ArgumentError("The 'db_index' parameter must be a Boolean"))
+  !(editable isa Bool) && throw(ArgumentError("The 'editable' parameter must be a Boolean"))
+  
+  return sBigIntegerField(
+    verbose_name,
+    false, # primary_key
+    unique,
+    blank,
+    null,
+    db_index,
+    default,
+    editable,
+    "BIGINT",
+    format_number_sql
+  )  
+end
+
+mutable struct sBooleanField <: PormGField
+  verbose_name::Union{String, Nothing}
+  primary_key::Bool
+  unique::Bool
+  blank::Bool
+  null::Bool
+  db_index::Bool
+  default::Union{Bool, Nothing}
+  editable::Bool
+  type::String
+  formater::Function
+end
+
+
+"""
+    BooleanField(; kwargs...)
+
+A field for storing boolean (true/false) values, equivalent to PostgreSQL's BOOLEAN columns.
+
+The `BooleanField` stores binary true/false values and is ideal for flags, switches, status indicators, and any field that represents a yes/no or on/off state. It maps directly to PostgreSQL's BOOLEAN type and Julia's Bool type.
+
+# Keyword Arguments
+- `verbose_name::Union{String, Nothing} = nothing`: A human-readable name for the field
+- `unique::Bool = false`: Whether values in this field must be unique (rarely used with booleans)
+- `blank::Bool = false`: Whether the field can be left blank in forms
+- `null::Bool = false`: Whether the database column can store NULL values
+- `db_index::Bool = false`: Whether to create a database index on this field
+- `default::Union{Bool, Nothing} = nothing`: Default value for the field (true or false)
+- `editable::Bool = false`: Whether the field should be editable in forms
+
+# Examples
+
+Basic boolean flags:
+```julia
+User = Models.Model(
+    _id = IDField(),
+    username = CharField(max_length=150),
+    is_active = BooleanField(default=true),
+    is_staff = BooleanField(default=false),
+    email_verified = BooleanField(default=false)
+)
+```
+
+# Boolean Values and Conversion
+The field handles various input formats:
+- **Julia Bool**: `true`, `false`
+- **Integers**: `1` (true), `0` (false)
+- **Strings**: `"true"`, `"false"`, `"1"`, `"0"`, `"yes"`, `"no"`
+- **NULL**: When `null=true`, accepts `NULL`/`nothing`
+"""
+function BooleanField(; kwargs...)
+  # List of accepted parameters
+  accepted = Set([
+      :verbose_name, :unique, :blank, :null, :db_index, :default, :editable
+  ])
+  # Check for unexpected parameters
+  for (k, v) in kwargs
+      if !(k in accepted)
+          @warn "Unexpected parameter for BooleanField. It will be ignored." field="BooleanField" param=k value=v
+      end
+  end
+  # Extract parameters with defaults
+  verbose_name = get(kwargs, :verbose_name, nothing)
+  unique = get(kwargs, :unique, false)
+  blank = get(kwargs, :blank, false)
+  null = get(kwargs, :null, false)
+  db_index = get(kwargs, :db_index, false)
+  default = get(kwargs, :default, nothing)
+  editable = get(kwargs, :editable, false)
+
+  # Validate verbose_name
+  !(verbose_name isa Union{Nothing, String}) && throw(ArgumentError("The 'verbose_name' must be a String or nothing"))
+  # Validate default
+  default = validate_default(default, Union{Bool, Nothing}, "BooleanField", x -> parse(Bool, string(x)))
+  # Validate other parameters
+  !(unique isa Bool) && throw(ArgumentError("The 'unique' must be a Boolean"))
+  !(blank isa Bool) && throw(ArgumentError("The 'blank' must be a Boolean"))
+  !(null isa Bool) && throw(ArgumentError("The 'null' must be a Boolean"))
+  !(db_index isa Bool) && throw(ArgumentError("The 'db_index' must be a Boolean"))
+  !(editable isa Bool) && throw(ArgumentError("The 'editable' must be a Boolean"))
+  # Return the field instance
+  return sBooleanField(
+    verbose_name,
+    false, # primary_key
+    unique,
+    blank,
+    null,
+    db_index,
+    default,
+    editable,
+    "BOOLEAN",
+    format_bool_sql
+  )  
+end
+
+mutable struct sDateField <: PormGField
+  verbose_name::Union{String, Nothing}
+  primary_key::Bool
+  unique::Bool
+  blank::Bool
+  null::Bool
+  db_index::Bool
+  default::Union{String, Nothing}
+  editable::Bool
+  auto_now::Bool
+  auto_now_add::Bool
+  type::String
+  formater::Function
+end
+
+"""
+    DateField(; kwargs...)
+
+A field for storing date values (without time), equivalent to PostgreSQL's DATE columns.
+
+The `DateField` stores calendar dates in YYYY-MM-DD format and is ideal for birth dates, event dates, deadlines, and any date information that doesn't require time precision. It maps to PostgreSQL's DATE type and Julia's Date type.
+
+# Keyword Arguments
+- `verbose_name::Union{String, Nothing} = nothing`: A human-readable name for the field
+- `unique::Bool = false`: Whether values in this field must be unique across all records
+- `blank::Bool = false`: Whether the field can be left blank in forms
+- `null::Bool = false`: Whether the database column can store NULL values
+- `db_index::Bool = false`: Whether to create a database index on this field
+- `default::Union{String, Nothing} = nothing`: Default value for the field (YYYY-MM-DD format)
+- `editable::Bool = false`: Whether the field should be editable in forms
+- `auto_now::Bool = false`: Whether to automatically set to current date on every save
+- `auto_now_add::Bool = false`: Whether to automatically set to current date on creation only
+
+# Examples
+
+Basic date fields:
+```julia
+User = Models.Model(
+    _id = IDField(),
+    username = CharField(max_length=150),
+    birth_date = DateField(null=true, blank=true),
+    join_date = DateField(auto_now_add=true),
+    last_login_date = DateField(null=true)
+)
+```
+
+Event and scheduling:
+```julia
+Event = Models.Model(
+    _id = IDField(),
+    title = CharField(max_length=200),
+    event_date = DateField(db_index=true),
+    registration_deadline = DateField(),
+    created_date = DateField(auto_now_add=true)
+)
+```
+
+Business dates:
+```julia
+Invoice = Models.Model(
+    _id = IDField(),
+    customer = ForeignKey("Customer"),
+    issue_date = DateField(auto_now_add=true),
+    due_date = DateField(),
+    paid_date = DateField(null=true, blank=true)
+)
+```
+
+# Auto Date Features
+
+## auto_now_add
+Sets the date automatically when the record is first created:
+```julia
+created_date = DateField(auto_now_add=true)
+# Automatically set to today's date on creation
+# Never changes after initial creation
+```
+
+## auto_now  
+Updates the date automatically every time the record is saved:
+```julia
+last_modified_date = DateField(auto_now=true)
+# Set to today's date on every save operation
+# Useful for tracking last update dates
+```
+
+# Date Input Formats
+The field accepts various input formats:
+- **Julia Date**: `Date(2024, 7, 28)`
+- **DateTime**: `DateTime(2024, 7, 28, 10, 30)` (time ignored)
+- **String ISO**: `"2024-07-28"`
+- **String formats**: Various date strings parseable by Julia
+"""
+function DateField(; kwargs...)
+  # List of accepted parameters
+  accepted = Set([
+      :verbose_name, :unique, :blank, :null, :db_index, :default, :editable, :auto_now, :auto_now_add
+  ])
+  # Check for unexpected parameters
+  for (k, v) in kwargs
+      if !(k in accepted)
+          @warn "Unexpected parameter for DateField. It will be ignored." field="DateField" param=k value=v
+      end
+  end
+  # Extract parameters with defaults
+  verbose_name = get(kwargs, :verbose_name, nothing)
+  unique = get(kwargs, :unique, false)
+  blank = get(kwargs, :blank, false)
+  null = get(kwargs, :null, false)
+  db_index = get(kwargs, :db_index, false)
+  default = get(kwargs, :default, nothing)
+  editable = get(kwargs, :editable, false)
+  auto_now = get(kwargs, :auto_now, false)
+  auto_now_add = get(kwargs, :auto_now_add, false)
+
+  # Validate verbose_name
+  !(verbose_name isa Union{Nothing, String}) && throw(ArgumentError("The verbose_name must be a String or nothing"))
+  # Validate default
+  default = validate_default(default, Union{Date, Nothing}, "DateField", format_date_sql)
+  # Validate other parameters
+  !(unique isa Bool) && throw(ArgumentError("The 'unique' must be a Boolean"))
+  !(blank isa Bool) && throw(ArgumentError("The 'blank' must be a Boolean"))
+  !(null isa Bool) && throw(ArgumentError("The 'null' must be a Boolean"))
+  !(db_index isa Bool) && throw(ArgumentError("The 'db_index' must be a Boolean"))
+  !(editable isa Bool) && throw(ArgumentError("The 'editable' must be a Boolean"))
+  !(auto_now isa Bool) && throw(ArgumentError("The 'auto_now' must be a Boolean"))
+  !(auto_now_add isa Bool) && throw(ArgumentError("The 'auto_now_add' must be a Boolean"))
+  # Return the field instance
+  return sDateField(
+    verbose_name,
+    false, # primary_key
+    unique,
+    blank,
+    null,
+    db_index,
+    default,
+    editable,
+    auto_now,
+    auto_now_add,
+    "DATE",
+    format_date_sql
+  )  
+end
+
+mutable struct sDateTimeField <: PormGField
+  verbose_name::Union{String, Nothing}
+  primary_key::Bool
+  unique::Bool
+  blank::Bool
+  null::Bool
+  db_index::Bool
+  default::Union{String, Nothing}
+  editable::Bool
+  auto_now::Bool
+  auto_now_add::Bool
+  type::String
+  formater::Function
+end
+
+"""
+    DateTimeField(; kwargs...)
+
+A field for storing date and time values with timezone information.
+
+# Keyword Arguments
+- `verbose_name::Union{String, Nothing}`: Human-readable name for the field. Default: `nothing`
+- `unique::Bool`: If `true`, ensures field values are unique across the table. Default: `false`
+- `blank::Bool`: If `true`, allows empty values in forms/validation. Default: `false`
+- `null::Bool`: If `true`, allows NULL values in the database. Default: `false`
+- `db_index::Bool`: If `true`, creates a database index for faster queries. Default: `false`
+- `default::Union{DateTime, Nothing, String}`: Default value for the field. Can be a DateTime object, ISO string, or `nothing`. Default: `nothing`
+- `editable::Bool`: If `true`, field can be edited in forms. Default: `false`
+- `auto_now::Bool`: If `true`, automatically updates to current datetime on every save. Default: `false`
+- `auto_now_add::Bool`: If `true`, automatically sets to current datetime when record is created. Default: `false`
+
+# Examples
+```julia
+# Basic datetime field
+created_at = DateTimeField()
+
+# Auto-timestamp fields
+created_at = DateTimeField(auto_now_add=true)
+updated_at = DateTimeField(auto_now=true)
+
+# Indexed datetime for queries
+event_time = DateTimeField(db_index=true, verbose_name="Event Timestamp")
+
+# With default value
+scheduled_at = DateTimeField(default=DateTime(2024, 1, 1, 12, 0, 0))
+
+# Optional datetime field
+deadline = DateTimeField(null=true, blank=true)```
+```
+"""
+function DateTimeField(; kwargs...)
+  # List of accepted parameters
+  accepted = Set([
+      :verbose_name, :unique, :blank, :null, :db_index, :default, :editable, :auto_now, :auto_now_add
+  ])
+  # Check for unexpected parameters
+  for (k, v) in kwargs
+      if !(k in accepted)
+          @warn "Unexpected parameter for DateTimeField. It will be ignored." field="DateTimeField" param=k value=v
+      end
+  end
+  # Extract parameters with defaults
+  verbose_name = get(kwargs, :verbose_name, nothing)
+  unique = get(kwargs, :unique, false)
+  blank = get(kwargs, :blank, false)
+  null = get(kwargs, :null, false)
+  db_index = get(kwargs, :db_index, false)
+  default = get(kwargs, :default, nothing)
+  editable = get(kwargs, :editable, false)
+  auto_now = get(kwargs, :auto_now, false)
+  auto_now_add = get(kwargs, :auto_now_add, false)
+
+  # Validate verbose_name
+  !(verbose_name isa Union{Nothing, String}) && throw(ArgumentError("The verbose_name must be a String or nothing"))
+  # Validate default
+  default = validate_default(default, Union{DateTime, Nothing}, "DateTimeField", x -> DateTime(x))
+  # Validate other parameters
+  !(unique isa Bool) && throw(ArgumentError("The 'unique' must be a Boolean"))
+  !(blank isa Bool) && throw(ArgumentError("The 'blank' must be a Boolean"))
+  !(null isa Bool) && throw(ArgumentError("The 'null' must be a Boolean"))
+  !(db_index isa Bool) && throw(ArgumentError("The 'db_index' must be a Boolean"))
+  !(editable isa Bool) && throw(ArgumentError("The 'editable' must be a Boolean"))
+  !(auto_now isa Bool) && throw(ArgumentError("The 'auto_now' must be a Boolean"))
+  !(auto_now_add isa Bool) && throw(ArgumentError("The 'auto_now_add' must be a Boolean"))
+  # Return the field instance
+  return sDateTimeField(
+    verbose_name,
+    false, # primary_key
+    unique,
+    blank,
+    null,
+    db_index,
+    default,
+    editable,
+    auto_now,
+    auto_now_add,
+    "TIMESTAMPTZ",
+    format_timezone_sql
+  )  
+end
+
+mutable struct sDecimalField <: PormGField
+  verbose_name::Union{String, Nothing}
+  primary_key::Bool
+  unique::Bool
+  blank::Bool
+  null::Bool
+  db_index::Bool
+  default::Union{Float64, Nothing}
+  editable::Bool
+  max_digits::Int
+  decimal_places::Int
+  type::String
+  formater::Function
+end
+
+"""
+    DecimalField(; kwargs...)
+
+A field for storing decimal numbers with fixed precision and scale.
+
+# Keyword Arguments
+- `verbose_name::Union{String, Nothing}`: Human-readable name for the field. Default: `nothing`
+- `unique::Bool`: If `true`, ensures field values are unique across the table. Default: `false`
+- `blank::Bool`: If `true`, allows empty values in forms/validation. Default: `false`
+- `null::Bool`: If `true`, allows NULL values in the database. Default: `false`
+- `db_index::Bool`: If `true`, creates a database index for faster queries. Default: `false`
+- `default::Union{Float64, Nothing}`: Default value for the field. Default: `nothing`
+- `editable::Bool`: If `true`, field can be edited in forms. Default: `false`
+- `max_digits::Int`: Maximum number of digits allowed (including decimal places). Default: `10`
+- `decimal_places::Int`: Number of decimal places to store. Default: `2`
+
+# Examples
+```julia
+# Currency field (2 decimal places)
+price = DecimalField(max_digits=10, decimal_places=2)
+
+# High precision scientific values
+measurement = DecimalField(max_digits=15, decimal_places=6)
+
+# Percentage with 4 decimal places
+rate = DecimalField(max_digits=7, decimal_places=4, default=0.0)
+
+# Financial calculation field
+amount = DecimalField(
+    max_digits=12, 
+    decimal_places=2, 
+    verbose_name="Transaction Amount",
+    db_index=true
+)
+
+# Optional decimal field
+discount = DecimalField(
+    max_digits=5, 
+    decimal_places=2, 
+    null=true, 
+    blank=true
+)
+```
+"""
+function DecimalField(; kwargs...)
+  # List of accepted parameters
+  accepted = Set([
+      :verbose_name, :unique, :blank, :null, :db_index, :default, :editable, :max_digits, :decimal_places
+  ])
+  # Check for unexpected parameters
+  for (k, v) in kwargs
+      if !(k in accepted)
+          @warn "Unexpected parameter for DecimalField. It will be ignored." field="DecimalField" param=k value=v
+      end
+  end
+  # Extract parameters with defaults
+  verbose_name = get(kwargs, :verbose_name, nothing)
+  unique = get(kwargs, :unique, false)
+  blank = get(kwargs, :blank, false)
+  null = get(kwargs, :null, false)
+  db_index = get(kwargs, :db_index, false)
+  default = get(kwargs, :default, nothing)
+  editable = get(kwargs, :editable, false)
+  max_digits = get(kwargs, :max_digits, 10)
+  decimal_places = get(kwargs, :decimal_places, 2)
+
+  # Validate verbose_name
+  !(verbose_name isa Union{Nothing, String}) && throw(ArgumentError("The verbose_name must be a String or nothing"))
+  
+  # Validate default using validate_default
+  default = validate_default(default, Union{Float64, Nothing}, "DecimalField", format2float64)
+  max_digits = validate_default(max_digits, Int, "DecimalField", format2int64)
+  decimal_places = validate_default(decimal_places, Int, "DecimalField", format2int64)
+  
+  # Validate other parameters
+  !(unique isa Bool) && throw(ArgumentError("The 'unique' parameter must be a Boolean"))
+  !(blank isa Bool) && throw(ArgumentError("The 'blank' parameter must be a Boolean"))
+  !(null isa Bool) && throw(ArgumentError("The 'null' parameter must be a Boolean"))
+  !(db_index isa Bool) && throw(ArgumentError("The 'db_index' parameter must be a Boolean"))
+  !(editable isa Bool) && throw(ArgumentError("The 'editable' parameter must be a Boolean"))
+  
+  return sDecimalField(
+    verbose_name,
+    false, # primary_key
+    unique,
+    blank,
+    null,
+    db_index,
+    default,
+    editable,
+    max_digits,
+    decimal_places,
+    "DECIMAL",
+    format_number_sql
+  )
+end
+
+mutable struct sEmailField <: PormGField
+  verbose_name::Union{String, Nothing}
+  primary_key::Bool
+  unique::Bool
+  blank::Bool
+  null::Bool
+  db_index::Bool
+  default::Union{String, Nothing}
+  editable::Bool
+  type::String
+  formater::Function
+end
+
+"""
+    EmailField(; kwargs...)
+
+A field for storing and validating email addresses.
+
+# Keyword Arguments
+- `verbose_name::Union{String, Nothing}`: Human-readable name for the field. Default: `nothing`
+- `unique::Bool`: If `true`, ensures field values are unique across the table. Default: `false`
+- `blank::Bool`: If `true`, allows empty values in forms/validation. Default: `false`
+- `null::Bool`: If `true`, allows NULL values in the database. Default: `false`
+- `db_index::Bool`: If `true`, creates a database index for faster queries. Default: `false`
+- `default::Union{String, Nothing}`: Default email address. Default: `nothing`
+- `editable::Bool`: If `true`, field can be edited in forms. Default: `false`
+
+# Examples
+```julia
+# Basic email field
+email = EmailField()
+
+# Unique email for user accounts
+user_email = EmailField(unique=true, verbose_name="User Email")
+
+# Optional contact email
+contact_email = EmailField(null=true, blank=true)
+
+# Email with default value
+notification_email = EmailField(default="admin@example.com")
+
+# Indexed email for fast lookups
+primary_email = EmailField(
+    unique=true, 
+    db_index=true, 
+)
+```
+"""
+function EmailField(; kwargs...)
+  # List of accepted parameters
+  accepted = Set([
+      :verbose_name, :unique, :blank, :null, :db_index, :default, :editable
+  ])
+  # Check for unexpected parameters
+  for (k, v) in kwargs
+      if !(k in accepted)
+          @warn "Unexpected parameter for EmailField. It will be ignored." field="EmailField" param=k value=v
+      end
+  end
+  # Extract parameters with defaults
+  verbose_name = get(kwargs, :verbose_name, nothing)
+  unique = get(kwargs, :unique, false)
+  blank = get(kwargs, :blank, false)
+  null = get(kwargs, :null, false)
+  db_index = get(kwargs, :db_index, false)
+  default = get(kwargs, :default, nothing)
+  editable = get(kwargs, :editable, false)
+
+  # Validate verbose_name
+  !(verbose_name isa Union{Nothing, String}) && throw(ArgumentError("The 'verbose_name' must be a String or nothing"))
+  # Validate default
+  default = validate_default(default, Union{String, Nothing}, "EmailField", x -> parse(String, x))
+  # Validate other parameters
+  !(unique isa Bool) && throw(ArgumentError("The 'unique' must be a Boolean"))
+  !(blank isa Bool) && throw(ArgumentError("The 'blank' must be a Boolean"))
+  !(null isa Bool) && throw(ArgumentError("The 'null' must be a Boolean"))
+  !(db_index isa Bool) && throw(ArgumentError("The 'db_index' must be a Boolean"))
+  !(editable isa Bool) && throw(ArgumentError("The 'editable' must be a Boolean"))
+  # Return the field instance
+  return sEmailField(
+    verbose_name,
+    false, # primary_key
+    unique,
+    blank,
+    null,
+    db_index,
+    default,
+    editable,
+    "VARCHAR",
+    format_text_sql
+  )  
+end
+
+# ============================================================================
+# Password Field 
+# ============================================================================
+
+# Password Field Definition
+
+mutable struct sPasswordField <: PormGField
+  verbose_name::Union{String, Nothing}
+  primary_key::Bool
+  unique::Bool
+  blank::Bool
+  null::Bool
+  db_index::Bool
+  default::Union{String, Nothing}
+  editable::Bool
+  type::String
+  formater::Function
+  max_length::Int  # Length of stored hash (Django uses VARCHAR(128))
+  auto_hash::Bool  # Whether to automatically hash on insert/update
+end
+
+"""
+    PasswordField(; kwargs...)
+
+A field for securely storing passwords using PBKDF2-SHA256 hashing (Django-compatible).
+
+The `PasswordField` stores hashed passwords in a format compatible with Django's authentication system. 
+Passwords are never stored in plain text - they are automatically hashed using PBKDF2-SHA256 with a 
+randomly generated salt and configurable iterations.
+
+# Keyword Arguments
+- `verbose_name::Union{String, Nothing} = nothing`: A human-readable name for the field
+- `blank::Bool = false`: Whether the field can be left blank in forms
+- `null::Bool = false`: Whether the database column can store NULL values  
+- `editable::Bool = true`: Whether the field should be editable in forms
+- `max_length::Int = 128`: Maximum length for stored hash (Django default)
+- `auto_hash::Bool = true`: Whether to automatically hash passwords on insert/update
+
+# Database Mapping
+- **PostgreSQL Type**: VARCHAR(128)
+- **Storage Format**: `pbkdf2_sha256\$iterations\$salt\$base64hash`
+- **Index**: Not indexed by default (passwords shouldn't be queried)
+
+# Stored Format
+The password is stored in Django-compatible format:
+```
+pbkdf2_sha256\$720000\$salt\$base64encodedHash
+```
+
+Where:
+- `pbkdf2_sha256`: Algorithm identifier
+- `720000`: Number of iterations
+- `salt`: Random 22-character salt
+- `base64encodedHash`: The derived key in base64
+
+# Examples
+
+Basic password field:
+```julia
+# Define a User model with password
+User = Models.Model(
+    _id = Models.IDField(),
+    username = Models.CharField(max_length=150, unique=true),
+    email = Models.EmailField(unique=true),
+    password = Models.PasswordField()
+)
+```
+
+Creating a user with hashed password:
+```julia
+import PormG.Models: make_password, check_password
+
+# Hash the password before storing
+hashed = make_password("securePassword123!")
+
+query = M.User |> object
+query.bulk_insert(
+    username = "ayrton_senna",
+    email = "senna@mclaren.com",
+    password = hashed
+)
+```
+
+Verifying a password during login:
+```julia
+import PormG.Models: check_password
+
+# Fetch user from database
+user_query = M.User |> object
+user_query.filter("username" => "ayrton_senna")
+users = user_query |> list
+
+if !isempty(users)
+    user = users[1]
+    # Verify the password
+    if check_password("securePassword123!", user[:password])
+        println("Login successful!")
+    else
+        println("Invalid password")
+    end
+end
+```
+
+# Related Functions
+- `make_password(raw)`: Hash a plain text password
+- `check_password(raw, encoded)`: Verify a password against stored hash
+- `password_needs_upgrade(encoded)`: Check if hash needs re-hashing
+
+# Security Notes
+- **Never store plain text passwords** - always use `make_password()`
+- The default 720000 iterations provides strong security (Django 4.2+ default)
+- Salts are automatically generated per-password
+- Use `check_password()` for verification (includes timing attack protection)
+- Consider implementing password upgrade on login if using older hashes
+
+# Migration from Django
+If you're migrating from a Django application, password hashes are fully compatible.
+Users can continue to log in without any password reset.
+
+# See Also
+- `CharField` for generic string storage
+- Django's password management documentation
+"""
+function PasswordField(; kwargs...)
+  # List of accepted parameters
+  accepted = Set([
+      :verbose_name, :blank, :null, :editable, :max_length, :auto_hash
+  ])
+  # Check for unexpected parameters
+  for (k, v) in kwargs
+      if !(k in accepted)
+          @warn "Unexpected parameter for PasswordField. It will be ignored." field="PasswordField" param=k value=v
+      end
+  end
+  # Extract parameters with defaults
+  verbose_name = get(kwargs, :verbose_name, nothing)
+  blank = get(kwargs, :blank, false)
+  null = get(kwargs, :null, false)
+  editable = get(kwargs, :editable, true)
+  max_length = get(kwargs, :max_length, 128)
+  auto_hash = get(kwargs, :auto_hash, true)
+
+  # Validate verbose_name
+  !(verbose_name isa Union{Nothing, String}) && throw(ArgumentError("The 'verbose_name' must be a String or nothing"))
+  # Validate other parameters
+  !(blank isa Bool) && throw(ArgumentError("The 'blank' must be a Boolean"))
+  !(null isa Bool) && throw(ArgumentError("The 'null' must be a Boolean"))
+  !(editable isa Bool) && throw(ArgumentError("The 'editable' must be a Boolean"))
+  !(max_length isa Int) && throw(ArgumentError("The 'max_length' must be an Integer"))
+  max_length < 64 && throw(ArgumentError("The 'max_length' must be at least 64 to store password hashes"))
+  !(auto_hash isa Bool) && throw(ArgumentError("The 'auto_hash' must be a Boolean"))
+  
+  # Return the field instance
+  return sPasswordField(
+    verbose_name,
+    false, # primary_key - passwords should never be primary keys
+    false, # unique - passwords should not be unique (allows same password for different users)
+    blank,
+    null,
+    false, # db_index - never index passwords
+    nothing, # default - no default password
+    editable,
+    "VARCHAR",
+    format_text_sql,
+    max_length,
+    auto_hash
+  )  
+end
+
+mutable struct sFloatField <: PormGField
+  verbose_name::Union{String, Nothing}
+  primary_key::Bool
+  unique::Bool
+  blank::Bool
+  null::Bool
+  db_index::Bool
+  default::Union{Float64, String, Int64, Nothing}
+  editable::Bool
+  type::String
+  formater::Function
+end
+
+"""
+    FloatField(; kwargs...)
+
+A field for storing floating-point numbers with double precision.
+
+# Keyword Arguments
+- `verbose_name::Union{String, Nothing}`: Human-readable name for the field. Default: `nothing`
+- `unique::Bool`: If `true`, ensures field values are unique across the table. Default: `false`
+- `blank::Bool`: If `true`, allows empty values in forms/validation. Default: `false`
+- `null::Bool`: If `true`, allows NULL values in the database. Default: `false`
+- `db_index::Bool`: If `true`, creates a database index for faster queries. Default: `false`
+- `default::Union{Float64, String, Int64, Nothing}`: Default value for the field. Default: `nothing`
+- `editable::Bool`: If `true`, field can be edited in forms. Default: `false`
+
+# Examples
+```julia
+# Basic float field
+temperature = FloatField()
+
+# Scientific measurement with default
+ph_level = FloatField(default=7.0)
+
+# Optional measurement
+weight = FloatField(null=true)
+```
+"""
+function FloatField(; kwargs...)
+  # List of accepted parameters
+  accepted = Set([
+      :verbose_name, :unique, :blank, :null, :db_index, :default, :editable
+  ])
+  # Check for unexpected parameters
+  for (k, v) in kwargs
+      if !(k in accepted)
+          @warn "Unexpected parameter for FloatField. It will be ignored." field="FloatField" param=k value=v
+      end
+  end
+  # Extract parameters with defaults
+  verbose_name = get(kwargs, :verbose_name, nothing)
+  unique = get(kwargs, :unique, false)
+  blank = get(kwargs, :blank, false)
+  null = get(kwargs, :null, false)
+  db_index = get(kwargs, :db_index, false)
+  default = get(kwargs, :default, nothing)
+  editable = get(kwargs, :editable, false)
+
+  # Validate verbose_name
+  !(verbose_name isa Union{Nothing, String}) && throw(ArgumentError("The verbose_name must be a String or nothing"))
+  
+  # Validate default using validate_default
+  default = validate_default(default, Union{Float64, String, Int64, Nothing}, "FloatField", parse)
+  
+  # Validate other parameters
+  !(unique isa Bool) && throw(ArgumentError("The 'unique' parameter must be a Boolean"))
+  !(blank isa Bool) && throw(ArgumentError("The 'blank' parameter must be a Boolean"))
+  !(null isa Bool) && throw(ArgumentError("The 'null' parameter must be a Boolean"))
+  !(db_index isa Bool) && throw(ArgumentError("The 'db_index' parameter must be a Boolean"))
+  !(editable isa Bool) && throw(ArgumentError("The 'editable' parameter must be a Boolean"))
+  
+  return sFloatField(
+    verbose_name,
+    false, # primary_key
+    unique,
+    blank,
+    null,
+    db_index,
+    default,
+    editable,
+    "FLOAT",
+    format_number_sql
+  )  
+end
+
+mutable struct sImageField <: PormGField
+  verbose_name::Union{String, Nothing}
+  primary_key::Bool
+  unique::Bool
+  blank::Bool
+  null::Bool
+  db_index::Bool
+  default::Union{String, Nothing}
+  editable::Bool
+  type::String
+  formater::Function
+end
+
+"""
+    ImageField(; kwargs...)
+
+A field for storing image file references and metadata.
+
+# Keyword Arguments
+- `verbose_name::Union{String, Nothing}`: Human-readable name for the field. Default: `nothing`
+- `unique::Bool`: If `true`, ensures field values are unique across the table. Default: `false`
+- `blank::Bool`: If `true`, allows empty values in forms/validation. Default: `false`
+- `null::Bool`: If `true`, allows NULL values in the database. Default: `false`
+- `db_index::Bool`: If `true`, creates a database index for faster queries. Default: `false`
+- `default::Union{String, Nothing}`: Default image path or URL. Default: `nothing`
+- `editable::Bool`: If `true`, field can be edited in forms. Default: `false`
+
+# Examples
+```julia
+# Basic image field
+avatar = ImageField()
+
+# Product image with default
+product_image = ImageField(
+    default="/static/images/default-product.jpg",
+    verbose_name="Product Image"
+)
+
+# Optional profile picture
+profile_pic = ImageField(null=true)
+
+# Unique banner image
+banner = ImageField(
+    unique=true,
+    db_index=true
+)
+```
+"""
+function ImageField(; kwargs...)
+  # List of accepted parameters
+  accepted = Set([
+      :verbose_name, :unique, :blank, :null, :db_index, :default, :editable
+  ])
+  # Check for unexpected parameters
+  for (k, v) in kwargs
+      if !(k in accepted)
+          @warn "Unexpected parameter for ImageField. It will be ignored." field="ImageField" param=k value=v
+      end
+  end
+  # Extract parameters with defaults
+  verbose_name = get(kwargs, :verbose_name, nothing)
+  unique = get(kwargs, :unique, false)
+  blank = get(kwargs, :blank, false)
+  null = get(kwargs, :null, false)
+  db_index = get(kwargs, :db_index, false)
+  default = get(kwargs, :default, nothing)
+  editable = get(kwargs, :editable, false)
+
+  # Validate verbose_name
+  !(verbose_name isa Union{Nothing, String}) && throw(ArgumentError("The 'verbose_name' must be a String or nothing"))
+  # Validate default
+  default = validate_default(default, Union{String, Nothing}, "ImageField", x -> parse(String, x))
+  # Validate other parameters
+  !(unique isa Bool) && throw(ArgumentError("The 'unique' must be a Boolean"))
+  !(blank isa Bool) && throw(ArgumentError("The 'blank' must be a Boolean"))
+  !(null isa Bool) && throw(ArgumentError("The 'null' must be a Boolean"))
+  !(db_index isa Bool) && throw(ArgumentError("The 'db_index' must be a Boolean"))
+  !(editable isa Bool) && throw(ArgumentError("The 'editable' must be a Boolean"))
+  # Return the field instance
+  return sImageField(
+    verbose_name,
+    false, # primary_key
+    unique,
+    blank,
+    null,
+    db_index,
+    default,
+    editable,
+    "BLOB",
+    format_text_sql
+  )  
+end
+
+mutable struct sTextField <: PormGField
+  verbose_name::Union{String, Nothing}
+  primary_key::Bool
+  unique::Bool
+  blank::Bool
+  null::Bool
+  db_index::Bool
+  default::Union{String, Nothing}
+  editable::Bool
+  type::String
+  formater::Function
+end
+
+"""
+    TextField(; kwargs...)
+
+A field for storing large amounts of text without length restrictions.
+
+# Keyword Arguments
+- `verbose_name::Union{String, Nothing}`: Human-readable name for the field. Default: `nothing`
+- `unique::Bool`: If `true`, ensures field values are unique across the table. Default: `false`
+- `blank::Bool`: If `true`, allows empty values in forms/validation. Default: `false`
+- `null::Bool`: If `true`, allows NULL values in the database. Default: `false`
+- `db_index::Bool`: If `true`, creates a database index for faster queries. Default: `false`
+- `default::Union{String, Nothing}`: Default text content. Default: `nothing`
+- `editable::Bool`: If `true`, field can be edited in forms. Default: `false`
+
+# Examples  
+```julia
+# Basic text field for long content
+description = TextField()
+
+# Blog post content
+content = TextField(blank=true)
+
+# Optional notes field
+notes = TextField(null=true, blank=true)
+
+# Indexed text field for search
+searchable_content = TextField(
+    db_index=true
+)
+
+# Text field with default content
+template = TextField(
+    default="Enter your text here...",
+    verbose_name="Template Content"
+)
+```
+"""
+function TextField(; kwargs...)
+  accepted = Set([:verbose_name, :unique, :blank, :null, :db_index, :default, :editable])
+  for (k, v) in kwargs
+      if !(k in accepted)
+          @warn "Unexpected parameter for TextField. It will be ignored." field="TextField" param=k value=v
+      end
+  end
+  verbose_name = get(kwargs, :verbose_name, nothing)
+  unique = get(kwargs, :unique, false)
+  blank = get(kwargs, :blank, false)
+  null = get(kwargs, :null, false)
+  db_index = get(kwargs, :db_index, false)
+  default = get(kwargs, :default, nothing)
+  editable = get(kwargs, :editable, false)
+
+  # Validate verbose_name
+  !(verbose_name isa Union{Nothing, String}) && throw(ArgumentError("The 'verbose_name' must be a String or nothing"))
+  # Validate default
+  default = validate_default(default, Union{String, Nothing}, "TextField", x -> parse(String, x))
+  # Validate other parameters
+  !(unique isa Bool) && throw(ArgumentError("The 'unique' must be a Boolean"))
+  !(blank isa Bool) && throw(ArgumentError("The 'blank' must be a Boolean"))
+  !(null isa Bool) && throw(ArgumentError("The 'null' must be a Boolean"))
+  !(db_index isa Bool) && throw(ArgumentError("The 'db_index' must be a Boolean"))
+  !(editable isa Bool) && throw(ArgumentError("The 'editable' must be a Boolean"))
+  # Return the field instance
+  return sTextField(
+    verbose_name,
+    false, # primary_key
+    unique,
+    blank,
+    null,
+    db_index,
+    default,
+    editable,
+    "TEXT",
+    format_text_sql
+  )  
+end
+
+mutable struct sTimeField <: PormGField
+  verbose_name::Union{String, Nothing}
+  primary_key::Bool
+  unique::Bool
+  blank::Bool
+  null::Bool
+  db_index::Bool
+  default::Union{String, Nothing}
+  editable::Bool
+  type::String
+  formater::Function
+end
+
+function TimeField(; kwargs...)
+  # List of accepted parameters
+  accepted = Set([
+      :verbose_name, :unique, :blank, :null, :db_index, :default, :editable
+  ])
+  # Check for unexpected parameters
+  for (k, v) in kwargs
+      if !(k in accepted)
+          @warn "Unexpected parameter for TimeField. It will be ignored." field="TimeField" param=k value=v
+      end
+  end
+  # Extract parameters with defaults
+  verbose_name = get(kwargs, :verbose_name, nothing)
+  unique = get(kwargs, :unique, false)
+  blank = get(kwargs, :blank, false)
+  null = get(kwargs, :null, false)
+  db_index = get(kwargs, :db_index, false)
+  default = get(kwargs, :default, nothing)
+  editable = get(kwargs, :editable, false)
+
+  # Validate verbose_name
+  !(verbose_name isa Union{Nothing, String}) && throw(ArgumentError("The 'verbose_name' must be a String or nothing"))
+  # Validate default
+  default = validate_default(default, Union{Time, Nothing}, "TimeField", x -> Time(x))
+  # Validate other parameters
+  !(unique isa Bool) && throw(ArgumentError("The 'unique' must be a Boolean"))
+  !(blank isa Bool) && throw(ArgumentError("The 'blank' must be a Boolean"))
+  !(null isa Bool) && throw(ArgumentError("The 'null' must be a Boolean"))
+  !(db_index isa Bool) && throw(ArgumentError("The 'db_index' must be a Boolean"))
+  !(editable isa Bool) && throw(ArgumentError("The 'editable' must be a Boolean"))
+  # Return the field instance
+  return sTimeField(
+    verbose_name,
+    false, # primary_key
+    unique,
+    blank,
+    null,
+    db_index,
+    default,
+    editable,
+    "TIME",
+    format_text_sql
+  )  
+end
+
+mutable struct sBinaryField <: PormGField
+  verbose_name::Union{String, Nothing}
+  primary_key::Bool
+  unique::Bool
+  blank::Bool
+  null::Bool
+  db_index::Bool
+  default::Union{String, Nothing}
+  editable::Bool
+  type::String
+  formater::Function
+  max_length::Union{Int, Nothing}
+end
+
+function BinaryField(; kwargs...)
+  # List of accepted parameters
+  accepted = Set([
+      :verbose_name, :unique, :blank, :null, :db_index, :default, :editable, :max_length
+  ])
+  # Check for unexpected parameters
+  for (k, v) in kwargs
+      if !(k in accepted)
+          @warn "Unexpected parameter for BinaryField. It will be ignored." field="BinaryField" param=k value=v
+      end
+  end
+  # Extract parameters with defaults
+  verbose_name = get(kwargs, :verbose_name, nothing)
+  unique = get(kwargs, :unique, false)
+  blank = get(kwargs, :blank, false)
+  null = get(kwargs, :null, false)
+  db_index = get(kwargs, :db_index, false)
+  default = get(kwargs, :default, nothing)
+  editable = get(kwargs, :editable, false)
+  max_length = get(kwargs, :max_length, nothing)
+
+  # Validate verbose_name
+  !(verbose_name isa Union{Nothing, String}) && throw(ArgumentError("The 'verbose_name' must be a String or nothing"))
+  # Validate default
+  default = validate_default(default, Union{Vector{UInt8}, Nothing}, "BinaryField", x -> Base64.decode(x))
+  if max_length isa AbstractString
+    if occursin(r"\d+", max_length)
+      max_length = validate_default(max_length, Int, "BinaryField", format2int64)
+    else
+      max_length = nothing
+    end
+  end
+  if !(max_length isa Union{Nothing, Int})
+    throw(ArgumentError("The 'max_length' must be an integer or nothing"))
+  elseif max_length isa Int && max_length <= 0
+    throw(ArgumentError("The 'max_length' must be a positive integer"))
+  end
+  # Validate other parameters
+  !(unique isa Bool) && throw(ArgumentError("The 'unique' must be a Boolean"))
+  !(blank isa Bool) && throw(ArgumentError("The 'blank' must be a Boolean"))
+  !(null isa Bool) && throw(ArgumentError("The 'null' must be a Boolean"))
+  !(db_index isa Bool) && throw(ArgumentError("The 'db_index' must be a Boolean"))
+  !(editable isa Bool) && throw(ArgumentError("The 'editable' must be a Boolean"))
+  # Return the field instance
+  return sBinaryField(
+    verbose_name,
+    false, # primary_key
+    unique,
+    blank,
+    null,
+    db_index,
+    default,
+    editable,
+    "BLOB",
+    format_text_sql,
+    max_length
+  )
+end
+
+mutable struct sDurationField <: PormGField
+  verbose_name::Union{String, Nothing}
+  primary_key::Bool
+  unique::Bool
+  blank::Bool
+  null::Bool
+  db_index::Bool
+  default::Union{Period, Nothing}
+  editable::Bool
+  type::String
+  formater::Function
+end
+
+function DurationField(; kwargs...)
+  # List of accepted parameters
+  accepted = Set([
+      :verbose_name, :unique, :blank, :null, :db_index, :default, :editable
+  ])
+  # Check for unexpected parameters
+  for (k, v) in kwargs
+      if !(k in accepted)
+          @warn "Unexpected parameter for DurationField. It will be ignored." field="DurationField" param=k value=v
+      end
+  end
+  # Extract parameters with defaults
+  verbose_name = get(kwargs, :verbose_name, nothing)
+  unique = get(kwargs, :unique, false)
+  blank = get(kwargs, :blank, false)
+  null = get(kwargs, :null, false)
+  db_index = get(kwargs, :db_index, false)
+  default = get(kwargs, :default, nothing)
+  editable = get(kwargs, :editable, false)
+
+  # Validate verbose_name
+  !(verbose_name isa Union{Nothing, String}) && throw(ArgumentError("The 'verbose_name' must be a String or nothing"))
+  # Validate default
+  default = validate_default(default, Union{Period, Nothing}, "DurationField", x -> parse(Period, string(x)))
+  # Validate other parameters
+  !(unique isa Bool) && throw(ArgumentError("The 'unique' must be a Boolean"))
+  !(blank isa Bool) && throw(ArgumentError("The 'blank' must be a Boolean"))
+  !(null isa Bool) && throw(ArgumentError("The 'null' must be a Boolean"))
+  !(db_index isa Bool) && throw(ArgumentError("The 'db_index' must be a Boolean"))
+  !(editable isa Bool) && throw(ArgumentError("The 'editable' must be a Boolean"))
+  # Return the field instance
+  return sDurationField(
+    verbose_name,
+    false, # primary_key
+    unique,
+    blank,
+    null,
+    db_index,
+    default,
+    editable,
+    "INTERVAL",
+    format_text_sql
+  )
+end
+
